Fourth homework(QA2). 1. Feature: Weather 2. Scenario: Checking coordinates for city

+++ src/test/java/model/Clouds.java
@@ -2,13 +2,13 @@ package model;
 
 public class Clouds {
 
-    private int all;
+    private long all;
 
-    public int getAll() {
+    public long getAll() {
         return all;
     }
 
-    public void setAll(int all) {
+    public void setAll(long all) {
         this.all = all;
     }
 }
